initialise slug generation when entry type changed and slug empty or temp
